Began Maze Generation
I wrote the code for the basic generation of the boundaries of the
maze, and then realized that the generation of the block used to create
the maze needed to be made in the runner class file, so I moved that. I
also started wrapping my head around the code involved with recursive
maze generation although this is not reflected in the code… yet.